Optimize Day 02 `is_safe_report_with_tolerance`
The function `check_all_variants_of_modification` no longer creates a copy of the vector to remove the level to ignore.
Instead, we check the original report but with additional parameters, which level has to be ignored.
Now, we don't need to allocate memory during the check, which makes the function 6 times faster.

```
day_02__count_safe_reports_with_tolerance
                        time:   [4.5785 µs 4.6083 µs 4.6355 µs]
                        change: [-84.538% -84.299% -84.105%] (p = 0.00 < 0.05)
                        Performance has improved.
```